+++ sdk/containerservice/azure-mgmt-containerservicefleet/azure/mgmt/containerservicefleet/aio/_configuration.py
@@ -34,8 +34,8 @@ class ContainerServiceFleetMgmtClientConfiguration:  # pylint: disable=too-many-
      None.
     :type cloud_setting: ~azure.core.AzureClouds
     :keyword api_version: The API version to use for this operation. Known values are
-     "2026-02-01-preview" and None. Default value is "2026-02-01-preview". Note that overriding this
-     default value may result in unsupported behavior.
+     "2025-08-01-preview". Default value is "2025-08-01-preview". Note that overriding this default
+     value may result in unsupported behavior.
     :paramtype api_version: str
     """
 
@@ -47,7 +47,7 @@ class ContainerServiceFleetMgmtClientConfiguration:  # pylint: disable=too-many-
         cloud_setting: Optional["AzureClouds"] = None,
         **kwargs: Any
     ) -> None:
-        api_version: str = kwargs.pop("api_version", "2026-02-01-preview")
+        api_version: str = kwargs.pop("api_version", "2025-08-01-preview")
 
         if credential is None:
             raise ValueError("Parameter 'credential' must not be None.")