+++ sdk/containerservice/azure-mgmt-containerservicefleet/generated_tests/test_container_service_fleet_mgmt_fleet_update_strategies_operations_async.py
@@ -56,10 +56,8 @@ class TestContainerServiceFleetMgmtFleetUpdateStrategiesOperationsAsync(AzureMgm
                                             "name": "str",
                                             "afterGates": [{"type": "str", "displayName": "str"}],
                                             "beforeGates": [{"type": "str", "displayName": "str"}],
-                                            "maxConcurrency": "str",
                                         }
                                     ],
-                                    "maxConcurrency": "str",
                                 }
                             ]
                         },